+++ src/components/ha-labels-picker.ts
@@ -17,9 +17,9 @@ import { showLabelDetailDialog } from "../panels/config/labels/show-dialog-label
 import type { HomeAssistant, ValueChangedEvent } from "../types";
 import "./chips/ha-chip-set";
 import "./chips/ha-input-chip";
+import type { HaDeviceComboBoxDeviceFilterFunc } from "./device/ha-device-picker";
 import "./ha-label-picker";
 import type { HaLabelPicker } from "./ha-label-picker";
-import type { HaDeviceComboBoxDeviceFilterFunc } from "./device/ha-device-combo-box";
 
 @customElement("ha-labels-picker")
 export class HaLabelsPicker extends SubscribeMixin(LitElement) {
