+++ src/components/ha-target-picker.ts
@@ -38,8 +38,8 @@ import { subscribeLabelRegistry } from "../data/label_registry";
 import { SubscribeMixin } from "../mixins/subscribe-mixin";
 import type { HomeAssistant } from "../types";
 import "./device/ha-device-combo-box";
-import type { HaDeviceComboBoxDeviceFilterFunc } from "./device/ha-device-combo-box";
 import "./device/ha-device-picker";
+import type { HaDeviceComboBoxDeviceFilterFunc } from "./device/ha-device-picker";
 import "./entity/ha-entity-picker";
 import type { HaEntityPickerEntityFilterFunc } from "./entity/ha-entity-picker";
 import "./ha-area-floor-picker";
